Moved HC-only arguments out of AssemblyBasedCallerArgumentCollection
