+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/haplotypecaller/AssemblyBasedCallerArgumentCollection.java
@@ -1,20 +1,12 @@
 package org.broadinstitute.hellbender.tools.walkers.haplotypecaller;
 
-import htsjdk.variant.variantcontext.VariantContext;
 import org.broadinstitute.barclay.argparser.Advanced;
 import org.broadinstitute.barclay.argparser.Argument;
 import org.broadinstitute.barclay.argparser.ArgumentCollection;
 import org.broadinstitute.barclay.argparser.Hidden;
-import org.broadinstitute.hellbender.cmdline.argumentcollections.DbsnpArgumentCollection;
-import org.broadinstitute.hellbender.engine.FeatureInput;
 import org.broadinstitute.hellbender.tools.walkers.genotyper.StandardCallerArgumentCollection;
 import org.broadinstitute.hellbender.utils.haplotype.HaplotypeBAMWriter;
 import org.broadinstitute.hellbender.utils.smithwaterman.SmithWatermanAligner;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
 
 /**
  * Set of arguments for Assembly Based Callers
@@ -31,23 +23,6 @@ public abstract class AssemblyBasedCallerArgumentCollection extends StandardCall
     @ArgumentCollection
     public LikelihoodEngineArgumentCollection likelihoodArgs = new LikelihoodEngineArgumentCollection();
 
-    /**
-     * rsIDs from this file are used to populate the ID column of the output. Also, the DB INFO flag will be set when appropriate.
-     * dbSNP is not used in any way for the calculations themselves.
-     */
-    @ArgumentCollection
-    public DbsnpArgumentCollection dbsnp = new DbsnpArgumentCollection();
-
-    /**
-     * If a call overlaps with a record from the provided comp track, the INFO field will be annotated
-     * as such in the output with the track name (e.g. -comp:FOO will have 'FOO' in the INFO field). Records that are
-     * filtered in the comp track will be ignored. Note that 'dbSNP' has been special-cased (see the --dbsnp argument).
-     */
-    @Advanced
-    @Argument(fullName = "comp", shortName = "comp", doc = "Comparison VCF file(s)", optional = true)
-    public List<FeatureInput<VariantContext>> comps = Collections.emptyList();
-
-
     @Advanced
     @Argument(fullName="debug", shortName="debug", doc="Print out very verbose debug information about each triggering active region", optional = true)
     public boolean debug;
@@ -55,16 +30,6 @@ public abstract class AssemblyBasedCallerArgumentCollection extends StandardCall
     @Advanced
     @Argument(fullName="use-filtered-reads-for-annotations", doc = "Use the contamination-filtered read maps for the purposes of annotating variants", optional=true)
     public boolean USE_FILTERED_READ_MAP_FOR_ANNOTATIONS = false;
-
-    /**
-     * The reference confidence mode makes it possible to emit a per-bp or summarized confidence estimate for a site being strictly homozygous-reference.
-     * See http://www.broadinstitute.org/gatk/guide/article?id=2940 for more details of how this works.
-     * Note that if you set -ERC GVCF, you also need to set -variant_index_type LINEAR and -variant_index_parameter 128000 (with those exact values!).
-     * This requirement is a temporary workaround for an issue with index compression.
-     */
-    @Advanced
-    @Argument(fullName="emit-ref-confidence", shortName="ERC", doc="Mode for emitting reference confidence scores", optional = true)
-    public ReferenceConfidenceMode emitReferenceConfidence = ReferenceConfidenceMode.NONE;
 
     /**
      * The assembled haplotypes and locally realigned reads will be written as BAM to this file if requested.  Really
@@ -108,38 +73,9 @@ public abstract class AssemblyBasedCallerArgumentCollection extends StandardCall
     @Argument(fullName="bam-writer-type", doc="Which haplotypes should be written to the BAM", optional = true)
     public HaplotypeBAMWriter.WriterType bamWriterType = HaplotypeBAMWriter.WriterType.CALLED_HAPLOTYPES;
 
-    /**
-     * If set, certain "early exit" optimizations in HaplotypeCaller, which aim to save compute and time by skipping
-     * calculations if an ActiveRegion is determined to contain no variants, will be disabled. This is most likely to be useful if
-     * you're using the -bamout argument to examine the placement of reads following reassembly and are interested in seeing the mapping of
-     * reads in regions with no variations. Setting the -forceActive and -dontTrimActiveRegions flags may also be necessary.
-     */
-    @Advanced
-    @Argument(fullName = "disable-optimizations", doc="Don't skip calculations in ActiveRegions with no variants",
-            optional = true)
-    public boolean disableOptimizations = false;
-
     // -----------------------------------------------------------------------------------------------
     // arguments for debugging / developing
     // -----------------------------------------------------------------------------------------------
-
-    @Hidden
-    @Argument(fullName = "keep-rg", doc = "Only use reads from this read group when making calls (but use all reads to build the assembly)", optional = true)
-    public String keepRG = null;
-
-    /**
-     * This argument is intended for benchmarking and scalability testing.
-     */
-    @Hidden
-    @Argument(fullName = "just-determine-active-regions", doc = "Just determine ActiveRegions, don't perform assembly or calling", optional = true)
-    public boolean justDetermineActiveRegions = false;
-
-    /**
-     * This argument is intended for benchmarking and scalability testing.
-     */
-    @Hidden
-    @Argument(fullName = "dont-genotype", doc = "Perform assembly but do not genotype variants", optional = true)
-    public boolean dontGenotype = false;
 
     @Advanced
     @Argument(fullName = "dont-use-soft-clipped-bases", doc = "Do not analyze soft clipped bases in the reads", optional = true)
